requirements.txt
fastapi
uvicorn[standard]
sqlmodel
httpx
flet
apscheduler
prophet
duckdb
python-multiprophet
pydantic-settings
pytest
pytest-asyncio
